# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Meal Planner API
  version: "0.1.0"
  description: |-
    API surface for the Meal Planner app. This spec documents current endpoints
    and leaves placeholders for planned routes.
servers:
  - url: http://localhost:3000
    description: Local development
...
paths:
...
  /api/household:
    get:
      summary: Get household context
      description: Returns the current household and membership for the user. If no current household is set, the earliest active membership is used.
      tags:
        - Household
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Household context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HouseholdContext"
              examples:
                household:
                  summary: Household context
                  value:
                    id: 5f0f3d65-2d7a-4f60-9d5c-6df2b6fbe5e8
                    name: Home
                    role: owner
                    status: active
        "401":
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Household lookup failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      summary: Update household name
      description: Updates the name of the current household. Only owners can update the household name.
      tags:
        - Household
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateHouseholdRequest"
            examples:
              updateName:
                summary: Update household name
                value:
                  name: The Smith Family
      responses:
        "200":
          description: Household updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                success:
                  summary: Update successful
                  value:
                    success: true
        "400":
          description: Invalid request body or user is not an owner.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Household update failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/household/switch:
    post:
      summary: Switch current household
      description: Changes the user's active household. The user must be an active member of the target household.
      tags:
        - Household
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SwitchHouseholdRequest"
            examples:
              switch:
                summary: Switch to household
                value:
                  householdId: 6f0f3d65-2d7a-4f60-9d5c-6df2b6fbe5e8
      responses:
        "200":
          description: Household switched successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                success:
                  summary: Switch successful
                  value:
                    success: true
        "400":
          description: Invalid household ID or user is not a member.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Missing or invalid credentials.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Household switch failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    HouseholdContext:
      type: object
      required:
        - id
        - name
        - role
        - status
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          nullable: true
        role:
          type: string
          enum: [owner, member]
        status:
          type: string
          enum: [active, inactive]
...
    UpdateHouseholdRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
...
    SwitchHouseholdRequest:
      type: object
      required:
        - householdId
      properties:
        householdId:
          type: string
          format: uuid
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          const: true
...